Freely sharing metadata
Comment:
As per email correspondence with Philippe Kahane of the University of
Grenoble, dated 2022-12-09, we can release this without restrictive
licensing. Olivier David has not replied for multiple weeks now. I would
suggest proceeding and notifying again via email. We are not releasing
any data which is not already in this repository, we are simply updating
the licensing to permit what has already been done with this repo the
entire time.