Inheritance Principle datasets: More metadata
Utilise tailored metadata key-values to highlight not only which metadata files are inherited from for any given data file, but also the value that should be inherited when multiple applicable key-value metadata files contain different values for the same key.
Also add dataset "ipi1195v2", which has the filesystem structure of the example shown in Issue #1195, but includes key-value metadata that highlights the ambiguity that arises if trying to permit metadata overloading when there is full flexibility in applicable metadata files at a single filesystem hierarchy level.